BackupIQ/requirements.txt
# Production Dependencies
# Core framework
aiofiles>=23.0.0
aiohttp>=3.8.0
asyncio-throttle>=1.0.0

# Configuration and validation
PyYAML>=6.0
jsonschema>=4.17.0
python-dotenv>=1.0.0

# Cloud providers
pyicloud>=1.0.0
google-cloud-storage>=2.7.0
boto3>=1.26.0

# Database
neo4j>=5.3.0
redis>=4.5.0
sqlalchemy>=1.4.46

# Monitoring and observability
prometheus-client>=0.16.0
structlog>=22.3.0
opentelemetry-api>=1.15.0
opentelemetry-instrumentation>=0.36b0

# Security
cryptography>=3.4.8
pyjwt>=2.6.0
bcrypt>=4.0.1

# Analysis and processing
numpy>=1.24.0
pandas>=1.5.0
scikit-learn>=1.2.0

# System monitoring
psutil>=5.9.0

# HTTP client
httpx>=0.23.0
requests>=2.28.0

# Utility
click>=8.1.0
rich>=13.0.0
typer>=0.7.0

# Data serialization
msgpack>=1.0.4
orjson>=3.8.0

# Async support
uvloop>=0.17.0; sys_platform != "win32"
asyncpg>=0.27.0

# File handling
pathlib2>=2.3.7
watchdog>=2.2.0

# Networking
certifi>=2022.12.7